update sample to use document get_words
